Preparing for version 1 changes
This essentially allows for an Auth pass-through whereby a user object (given by the auth manager) is used to request a new token, rather than the current method where a user's credentials are needed to then get that user and then issue a token.

This allows the user to authenticated elsewhere and by whatever means is most suitable (i.e. using a Basic auth filter) and means this package doesn't need to care about authenticating a user, just giving them a token.

For version 1.0 this will become the default method of operation (currently the package provides a backwards compatible fix for existing implementations).

The other changes that we should make for version 1 is adding a user model and tokens controller and documenting the user of alternative implementations.